Compiler/Parser: Add default functionality & start new node.
When a nodes `node.location()` has not been overriden the program just
exits with info.

It's not very great info right now.

This also adds a scaffold for a `assignment_statement` which is a
declaration of a variable.

For example, this:
`unsigned long a = 2`
is an `assignment_statement`

Whereas:
`a = 4` is an `assignment`